Fix bug #1586962; EDIT no longer displays editor when given bad input.
Fixing this bug was a simple matter of adding a NOT_THROWING check.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@1755 9d0050d7-00b0-40f9-871c-8e02cbd5794f

+++ trunk/logo/src/wrksp.cpp
@@ -1378,15 +1378,20 @@ NODE *ledit(NODE *args)
          }
       }
 
-   NODE * args_list = vref(args);
-   if (TMainFrame::PopupEditorForFile(TempPathName, args_list))
+   if (NOT_THROWING)
       {
-      err_logo(FILE_ERROR, make_static_strnode("Could not launch the editor"));
-      }
-   else
-      {
-      unlink(TempPathName);
-      IsDirty = true;
+      NODE * args_list = vref(args);
+      if (TMainFrame::PopupEditorForFile(TempPathName, args_list))
+         {
+         err_logo(
+            FILE_ERROR, 
+            make_static_strnode("Could not launch the editor"));
+         }
+      else
+         {
+         unlink(TempPathName);
+         IsDirty = true;
+         }
       }
 
    lsetcursorarrow(NIL);
